requirements.txt
fire
langchain
python-dotenv
zhipuai
bs4
html2text